DELETE Classroom and User disabled working

+++ frontend/Web-Angular/Gepeto/src/app/view/classrooms/classrooms.component.ts
@@ -57,46 +57,41 @@ export class ClassroomsComponent implements OnInit {
   removeClassroom(classroomRemoved: Classroom) { //aqui é a SP_DELETE_CLASSROOM
     // const idClassroomToBeRemovedString = JSON.stringify(new Number(classroomRemoved[0]))
     // const idClassroomToBeRemovedNumber: number = parseInt(idClassroomToBeRemovedString)
-    // const Groups = this.GroupService.showGroups(classroomRemoved["idClassroom"]);
-    // Groups.subscribe(data => {
-    //   if(data.length == 0){
-    //     alert('Esta sala não pode ser deletada')
-    //   }
-    // }
-    //                      )
-    this.ClassroomService.removeClassroom(classroomRemoved["idClassroom"])
-      .subscribe(
-        result => {
-          this.GroupService.showGroups
-          alert('Sala "' + classroomRemoved["nameClassroom"] + '" Excluida com Sucesso')
-          const index = this.classrooms.indexOf(classroomRemoved)
-          this.classrooms.splice(index, 1)
-        },
-        error => {
-          if (error.status == 500) {
-            alert('Essa sala não pode ser removida, pois já contem grupos.')
-          }
-        }
-      )
+    const Groups = this.GroupService.showGroups(classroomRemoved["idClassroom"]);
+    Groups.subscribe(data => {
+      if (data.length != 0) {
+        alert('Esta sala não pode ser deletada')
+      }
+      else {
+        this.ClassroomService.removeClassroom(classroomRemoved["idClassroom"])
+          .subscribe(
+            result => {
+              this.GroupService.showGroups
+              alert('Sala "' + classroomRemoved["nameClassroom"] + '" Excluida com Sucesso')
+              const index = this.classrooms.indexOf(classroomRemoved)
+              this.classrooms.splice(index, 1)
+            })
+      }
+    })
   }
 
   showClassrooms() {
-    this.ClassroomService.showClassrooms()
-      .subscribe(data => this.classrooms = data);
-  }
+      this.ClassroomService.showClassrooms()
+        .subscribe(data => this.classrooms = data);
+    }
   // changeGroupsUrlService(classroomId: number) {
   //   // this.GroupService.showGroupsUrl = this.GroupService.showGroupsUrl.replace(/\d+/g, '')
   //   // this.GroupService.showGroupsUrl = this.GroupService.showGroupsUrl + classroomId
   //   this.GroupService.showGroupsUrl = this.GroupService.showGroupsUrl.substring(0,17)
   //   this.GroupService.changeCurrentClassroom(classroomId);
   // }
-  clearLocal(){
-    localStorage.clear();
-  }
-  changeCurrentClassroom(classroomId: number){
-    localStorage.removeItem('currentClassroom')
+  clearLocal() {
+      localStorage.clear();
+    }
+  changeCurrentClassroom(classroomId: number) {
+      localStorage.removeItem('currentClassroom')
     const classroomIdString = classroomId.toString()
     localStorage.setItem('currentClassroom', classroomIdString)
-  }
+    }
 
 }
